Move ImageIO logic from raylib.cc to image.cc
